restructure components directory and change title and favicons
