+++ Frontend/src/App.jsx
@@ -1,11 +1,11 @@
 import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Layout from "./components/Layout/LandingPage/Layout";
-import NotFound from "./components/Pages/NotFound/NotFound";
-import HomePage from "./components/Pages/LandingPage/HomePage";
-import ContactUs from "./components/Pages/LandingPage/ContactUs";
-import Login from "./components/Pages/Authentication/Login";
-import SignUp from "./components/Pages/Authentication/SignUp";
+import NotFound from "./Pages/NotFound/NotFound";
+import HomePage from "./Pages/LandingPage/HomePage";
+import ContactUs from "./Pages/LandingPage/ContactUs";
+import Login from "./Pages/Authentication/Login";
+import SignUp from "./Pages/Authentication/SignUp";
 
 function App() {
   return (
