+++ Frontend/src/Pages/LandingPage/ContactUs.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import ContactForm from '../../components/LandingPage/ContactUs/ContactForm.jsx'
+import CallToAction from '../../components/LandingPage/ContactUs/CallToAction'
+import Faq from '../../components/LandingPage/ContactUs/Faq.jsx'
+
+const ContactUs = () => {
+  return (
+    <>
+      <ContactForm />
+      <CallToAction />
+      <Faq />
+    </>
+  )
+}
+
+export default ContactUs
