--- Frontend/src/components/Pages/LandingPage/ContactUs.jsx
@@ -1,16 +1,0 @@
-import React from 'react'
-import ContactForm from '../../LandingPage/ContactUs/ContactForm.jsx'
-import CallToAction from '../../LandingPage/ContactUs/CallToAction'
-import Faq from '../../LandingPage/ContactUs/Faq.jsx'
-
-const ContactUs = () => {
-  return (
-    <>
-      <ContactForm />
-      <CallToAction />
-      <Faq />
-    </>
-  )
-}
-
-export default ContactUs
