--- Frontend/src/components/Pages/LandingPage/HomePage.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import Hero from  '../..//LandingPage/Home/Hero.jsx'
-import HowItWorks from '../../LandingPage/Home/HowItWorks.jsx';
-import CallToAction from '../../LandingPage/Home/CallToAction.jsx'
-
-const HomePage = () => {
-  return (
-    <div>
-        <Hero />
-        <HowItWorks />
-        <CallToAction />
-    </div>
-  )
-}
-
-export default HomePage
